% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tproperties.R
\name{buildname}
\alias{buildname}
\title{Build file name for saving figures or info}
\usage{
buildname(
  fig_or_info,
  analysis,
  variable = NA,
  type = NA,
  info = NA,
  period = NA
)
}
\arguments{
\item{fig_or_info}{type of file: "Fig" or "Info"}

\item{analysis}{type of analysis ("CWT" or "DWT" for instance)}

\item{variable}{name of variable ("y1" for instance)}

\item{type}{type of representation (1, 2 or 3)}

\item{info}{the name of info (coefficients)}

\item{period}{(the period if type=2)}
}
\description{
Build file name for saving figures or info
}
