% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cwt.R
\name{cwt_series_panel_ui}
\alias{cwt_series_panel_ui}
\title{Generate a CWT series panel}
\usage{
cwt_series_panel_ui(ns, series_id = 1)
}
\arguments{
\item{ns}{A namespace function (from `NS(id)`) for module namespacing.}

\item{series_id}{Integer, either 1 or 2, specifying which series panel to create.}
}
\value{
A `shiny.tag.list` containing the fluidRow for the specified CWT series.
}
\description{
This function creates a fluidRow containing a wellPanel for a CWT series
(either y1 or y2) with selectable plot type, sliders, conditional inputs,
hover outputs, and download buttons. The function adjusts all input/output
IDs according to the specified series number.
}
\examples{
ns <- NS("cwt")
cwt_series_panel_ui(ns, 1)
cwt_series_panel_ui(ns, 2)
}
