% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data.R
\name{data_series_panel_ui}
\alias{data_series_panel_ui}
\title{Generate a data (simulation) series panel}
\usage{
data_series_panel_ui(ns, series_id = 1)
}
\arguments{
\item{ns}{A namespace function (from `NS(id)`) for module namespacing.}

\item{series_id}{Integer, either 1 or 2, specifying which series panel to create.}
}
\value{
A `shiny.tag.list` containing the fluidRow for the specified data series.
}
\description{
This function creates a fluidRow containing a wellPanel for a data series
(either y1 or y2) with selectable plot type, sliders, conditional inputs,
hover outputs, and download buttons. The function adjusts all input/output
IDs according to the specified series number.
}
\examples{
ns <- NS("data")
data_series_panel_ui(ns, 1)
data_series_panel_ui(ns, 2)
}
