% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dwt.R
\name{dwt_series_panel_ui}
\alias{dwt_series_panel_ui}
\title{Generate a DWT series panel}
\usage{
dwt_series_panel_ui(ns, series_id = 1)
}
\arguments{
\item{ns}{A namespace function (from `NS(id)`) for module namespacing.}

\item{series_id}{Integer, either 1 or 2, specifying which series panel to create.}
}
\value{
A `shiny.tag.list` containing the fluidRow for the specified DWT series.
}
\description{
This function creates a fluidRow containing a wellPanel for a DWT series
(either y1 or y2) with selectable plot type, wavelet filter, conditional inputs,
slider, hover outputs, download buttons, and the plot. All input/output IDs
are adjusted according to the specified series number.
}
\examples{
ns <- NS("dwt")
dwt_series_panel_ui(ns, 1)
dwt_series_panel_ui(ns, 2)
}
