% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_cwt.R
\name{extract_local_extrema}
\alias{extract_local_extrema}
\title{Extract local extrema from a numeric vector}
\usage{
extract_local_extrema(y, minima_only = F, maxima_only = F)
}
\arguments{
\item{y}{Numeric vector.}

\item{minima_only}{Logical, default FALSE. If TRUE, only minima are returned.}

\item{maxima_only}{Logical, default FALSE. If TRUE, only maxima are returned.}
}
\value{
Integer vector of indices of local extrema.
}
\description{
Returns the indices of local maxima or minima.
}
\examples{
x <- c(1,3,2,4,3)
extract_local_extrema(x)          # indices of all extrema
extract_local_extrema(x, maxima_only=TRUE)
}
