% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findmypath.R
\name{findmypath}
\alias{findmypath}
\title{Find path to waveleT package}
\usage{
findmypath(dir, file)
}
\description{
This function is used internally to find package's path.
}
\keyword{toolkat}
\keyword{waveleT,}
\keyword{wavelet,}

