% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils.R
\name{format_time}
\alias{format_time}
\title{Formats time as a numeric vector (for use with wavelet transforms function) and step size}
\usage{
format_time(x, time_format)
}
\arguments{
\item{x}{the time vector}

\item{time_format}{the format of the time vector (as in `as.POSIXct` function)}
}
\value{
A list with original time vector(x) and derived numeric vector (xt), as well as original step size (step) and numerical step (stept.
}
\description{
Formats time as a numeric vector (for use with wavelet transforms function) and step size
}
\examples{
data(cat_moods)
format_time(cat_moods$x,"\%Y-\%m-\%d \%H:\%M:\%S")
}
