% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils.R
\name{makeDownloadHandler}
\alias{makeDownloadHandler}
\title{Crée un downloadHandler générique}
\usage{
makeDownloadHandler(
  id,
  fig_or_info,
  options = app_options(),
  input,
  content_fun,
  analysis,
  variable = NA,
  type = NA,
  info = NA,
  period = NA,
  ...
)
}
\arguments{
\item{id}{Integer. Identifiant (souvent 1 ou 2 dans les modules).}

\item{fig_or_info}{"Fig" or "Info}

\item{input}{L'objet `input` de Shiny.}

\item{content_fun}{Fonction prenant `(file, input, id, ...)` et écrivant le fichier.}

\item{...}{Arguments supplémentaires transmis aux deux fonctions `filename_fun` et `content_fun`.}

\item{filename_fun}{Fonction prenant `(input, id, ...)` et retournant un nom de fichier.}
}
\value{
Un objet de type `downloadHandler`.
}
\description{
Cette fonction fournit un wrapper générique autour de `downloadHandler`.
Elle permet de factoriser totalement la construction des handlers en séparant
la logique de génération du nom de fichier et la logique d'écriture du fichier.
}
