% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cwt.R
\name{mod_cwt_ui}
\alias{mod_cwt_ui}
\title{Module CWT (Continuous Wavelet Transform)
UI}
\usage{
mod_cwt_ui(id, app_data)
}
\arguments{
\item{id}{identifiant du module}

\item{app_data}{app_data}
}
\description{
Module CWT (Continuous Wavelet Transform)
UI
}
