% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dwt.R
\name{mod_dwt_ui}
\alias{mod_dwt_ui}
\title{Module DWT (Discrete Wavelet Transform)
UI}
\usage{
mod_dwt_ui(id)
}
\arguments{
\item{id}{identifiant du module}
}
\description{
Module DWT (Discrete Wavelet Transform)
UI
}
