% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_xwt.R
\name{mod_xwt_ui}
\alias{mod_xwt_ui}
\title{Module DWT (Discrete Wavelet Transform)
UI}
\usage{
mod_xwt_ui(id)
}
\arguments{
\item{id}{identifiant du module}
}
\description{
Module DWT (Discrete Wavelet Transform)
UI
}
