% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils.R
\name{numeric_step}
\alias{numeric_step}
\title{Extract numeric step from step string}
\usage{
numeric_step(step)
}
\arguments{
\item{step}{A string representing the step size, e.g. "3 hours".}
}
\value{
Numeric value of the step size.
}
\description{
Extract numeric step from step string
}
\examples{
waveleT:::numeric_step("10 mins")
}
