% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_dwt.R
\name{plotDWT_sig}
\alias{plotDWT_sig}
\title{Plot a single DWT/MRA signal with significance boundaries}
\usage{
plotDWT_sig(mra_sum, yscale = TRUE, plot = TRUE, col = "blue", xlim = NULL)
}
\arguments{
\item{mra_sum}{List output from `wav_mra_sum`, containing $y, $lab, $n.boundary.}

\item{yscale}{Logical, whether to use the original signal's y-scale.}

\item{plot}{Logical, whether to overlay the original signal.}

\item{col}{Color for the MRA sum line (default `"blue"`).}

\item{xlim}{Numeric vector, x-axis limits.}
}
\value{
Invisibly returns NULL. The function produces a plot.
}
\description{
Plot a single DWT/MRA signal with significance boundaries
}
\examples{
data(data_waveleT)
mra_obj=wav_mra_obj(data_waveleT,"x","y1",wt_filter="la8")
mra_data=wav_mra_data(mra_obj)
mra_sum <- wav_mra_sum(mra_data, c("D2","D3","S3"),wt_filter="la8")
plotDWT_sig(mra_sum, col="blue")
}
