% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_dwt.R
\name{plotDWT_superpose}
\alias{plotDWT_superpose}
\title{Superpose two DWT/MRA signals
Plots two MRA signals together, optionally overlaying the original signals.}
\usage{
plotDWT_superpose(
  mra_sum1,
  mra_sum2,
  yscale1 = TRUE,
  yscale2 = TRUE,
  plot1 = TRUE,
  plot2 = TRUE,
  xlim = NULL
)
}
\arguments{
\item{mra_sum1}{List output from `wav_mra_sum()` for the first signal.}

\item{mra_sum2}{List output from `wav_mra_sum()` for the second signal.}

\item{yscale1}{Logical, whether to use original signal y-scale for signal 1.}

\item{yscale2}{Logical, whether to use original signal y-scale for signal 2.}

\item{plot1}{Logical, whether to overlay original signal 1.}

\item{plot2}{Logical, whether to overlay original signal 2.}

\item{xlim}{Numeric vector, x-axis limits.}

\item{xname}{name of x variable}

\item{y1name}{name of y1 variable}

\item{y2name}{name of y2 variable}
}
\value{
Invisibly returns NULL. The function produces a plot.
}
\description{
Superpose two DWT/MRA signals
Plots two MRA signals together, optionally overlaying the original signals.
}
\examples{
data(data_waveleT)
mra_obj1=wav_mra_obj(data_waveleT,"x","y1",wt_filter="la8")
mra_obj2=wav_mra_obj(data_waveleT,"x","y2",wt_filter="la8")
mra_data1=wav_mra_data(mra_obj1)
mra_data2=wav_mra_data(mra_obj2)
mra_sum1 <- wav_mra_sum(mra_data1, c("D6","D7","S7"),wt_filter="la8")
mra_sum2 <- wav_mra_sum(mra_data2, c("D6","S6"),wt_filter="la8")
plotDWT_superpose(mra_sum1, mra_sum2)
}
