% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_cwt.R
\name{plotWT}
\alias{plotWT}
\title{Plot WT transforms results}
\usage{
plotWT(
  mywt,
  data,
  plot.sig,
  alpha,
  info,
  mother,
  plot.type,
  myperiod,
  plot.maxima,
  plot.minima,
  step,
  x_is_date,
  plot.cb,
  xlim,
  ylim,
  plot.phase
)
}
\arguments{
\item{mywt}{Wavelet object.}

\item{data}{Data matrix.}

\item{plot.sig}{Logical.}

\item{alpha}{Numeric.}

\item{info}{Character info type.}

\item{mother}{Character, mother wavelet.}

\item{plot.type}{Character, one of "1) Info=f(x,T)", "2) Info=f(x)", "3) Power=f(T)".}

\item{myperiod}{Numeric, selected period.}

\item{plot.maxima}{Logical.}

\item{plot.minima}{Logical.}

\item{step}{Numeric.}

\item{x_is_date}{Logical.}

\item{plot.cb}{Logical.}

\item{xlim}{Numeric vector.}

\item{ylim}{Numeric vector.}

\item{plot.phase}{Logical.}
}
\value{
Data frame of info used for plotting.
}
\description{
Plot WT transforms results
}
