% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_cwt.R
\name{plotWT_type1}
\alias{plotWT_type1}
\title{Plot wavelet analysis type 1 (Info=f(x,T))}
\usage{
plotWT_type1(
  data,
  mywt,
  plot.sig,
  info,
  step,
  units,
  plot.cb,
  xlim = NULL,
  ylim = NULL,
  plot.phase
)
}
\arguments{
\item{data}{Data matrix used for the wavelet analysis.}

\item{mywt}{Wavelet object.}

\item{plot.sig}{Logical, whether to plot significance.}

\item{info}{Character, type of info to plot.}

\item{step}{Numeric, step for scales.}

\item{units}{Character, units for time axis.}

\item{plot.cb}{Logical, whether to show colorbar.}

\item{xlim}{Numeric vector for x-axis limits.}

\item{ylim}{Numeric vector for y-axis limits.}

\item{plot.phase}{Logical, whether to plot phase arrows.}
}
\value{
Data frame of extracted info for the plotted wavelet.
}
\description{
Plot wavelet analysis type 1 (Info=f(x,T))
}
\examples{
\dontrun{
plotWT_type1(data_matrix, wt_obj, TRUE, "power", 1, "s", TRUE, NULL, NULL, FALSE)
}
}
