% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_cwt.R
\name{plotWT_type2}
\alias{plotWT_type2}
\title{#' Plot wavelet analysis type 2 (Info=f(x) at specific period)}
\usage{
plotWT_type2(
  data,
  mywt,
  info,
  myperiod,
  plot.sig,
  plot.maxima,
  plot.minima,
  step,
  units,
  xlim
)
}
\arguments{
\item{data}{Data matrix.}

\item{mywt}{Wavelet object.}

\item{info}{Character, info type.}

\item{myperiod}{Numeric, selected period.}

\item{plot.sig}{Logical, whether to plot significance.}

\item{plot.maxima}{Logical, whether to mark local maxima.}

\item{plot.minima}{Logical, whether to mark local minima.}

\item{step}{Numeric step.}

\item{units}{Character units.}

\item{xlim}{Numeric vector for x-axis limits.}
}
\value{
Data frame of extracted info at the specified period.
}
\description{
#' Plot wavelet analysis type 2 (Info=f(x) at specific period)
}
