% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_cwt.R
\name{plotWT_type3}
\alias{plotWT_type3}
\title{Plot wavelet analysis type 3 (Power=f(T))}
\usage{
plotWT_type3(data, mywt, plot.sig, alpha, step, units, ylim)
}
\arguments{
\item{data}{Data matrix.}

\item{mywt}{Wavelet object.}

\item{plot.sig}{Logical.}

\item{alpha}{Numeric significance level.}

\item{step}{Numeric step.}

\item{units}{Character units.}

\item{ylim}{Numeric vector for y-axis limits.}
}
\value{
Data frame of global power spectrum.
}
\description{
Plot wavelet analysis type 3 (Power=f(T))
}
