% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tproperties.R, R/fct_utils.R
\name{Tproperties}
\alias{Tproperties}
\title{Get info regarding x if x is time or date values}
\usage{
Tproperties(x, origin = x[1])

Tproperties(x, origin = x[1])
}
\description{
' @param x Numeric vector of time or date values.

' @param x Numeric vector of time or date values.
}
