% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_dwt.R
\name{wav_jmax}
\alias{wav_jmax}
\title{Maximum level of decomposition in DWT/MRA}
\usage{
wav_jmax(n, wt_filter = "la8", modwt = TRUE)
}
\arguments{
\item{n}{signal length}

\item{wt_filter}{Wavelet filter name (from wavelets' package `wt.filter`). Useful only if modwt=FALSE. Defaults to "la8".}

\item{modwt}{whether to consider MODWT (TRUE, default) or DWT (FALSE)}
}
\value{
An integer indicating the maximum level of decomposition
}
\description{
Maximum level of decomposition in DWT/MRA
}
\examples{
wav_jmax(1000)
wav_jmax(1000,wt_filter="la8",modwt=FALSE)
wav_jmax(1000,wt_filter="d4",modwt=FALSE)
}
