% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_dwt.R
\name{wav_mra_data}
\alias{wav_mra_data}
\title{Multi-resolution analysis (MRA) data.frame}
\usage{
wav_mra_data(mra_obj)
}
\arguments{
\item{mra_obj}{mra analysis as returned by wav_mra_obj()}
}
\value{
A data.frame with all detail and smooth sub-signals
}
\description{
Multi-resolution analysis (MRA) data.frame
}
\examples{
data(data_waveleT)
mra_obj=wav_mra_obj(data_waveleT,"x","y1",wt_filter="la8")
mra_data=wav_mra_data(mra_obj)
}
