% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_dwt.R
\name{wav_mra_obj}
\alias{wav_mra_obj}
\title{Multi-resolution analysis (MRA) object}
\usage{
wav_mra_obj(data, xname = "x", signalname, wt_filter = "la8")
}
\arguments{
\item{data}{Data frame with raw data}

\item{xname}{The name of the x (spatial or temporal position) column in data. Defaults to x.}

\item{signalname}{The name of the signal column in data}

\item{wt_filter}{Wavelet filter name (from wavelets' package `wt.filter`)}
}
\description{
Multi-resolution analysis (MRA) object
}
\examples{
data(data_waveleT)
mra_obj=wav_mra_obj(data_waveleT,"y1",wt_filter="la8")
}
