% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waveleT.R
\name{waveleT}
\alias{waveleT}
\title{Launch the Wavelet ToolKat}
\usage{
waveleT()
}
\description{
This function allows you to launch the Wavelet ToolKat interface. This is a Shiny App designed to help users analyze series through wavelet transforms.
For more information see package's vignette.
}
\examples{
waveleT()
}
\keyword{toolkat}
\keyword{waveleT,}
\keyword{wavelet,}

