% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_cwt.R
\name{wt_infomatrix}
\alias{wt_infomatrix}
\title{Extract information matrix from a wavelet object}
\usage{
wt_infomatrix(mywt, info)
}
\arguments{
\item{mywt}{A wavelet object, typically returned by `wt()`.}

\item{info}{Character. Type of information to extract: `"rsq"`, `"power"`, `"power.norm"`, `"power.corr"`, `"power.corr.norm"`, `"wavelet"`, `"phase"`.}
}
\value{
A matrix corresponding to the requested information.
}
\description{
Returns a matrix of values from a wavelet object based on the `info` argument.
}
\examples{
\dontrun{
wt_obj <- wt(matrix(rnorm(100), ncol=2))
M <- wt_infomatrix(wt_obj, "power")
}
}
