% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_cwt.R
\name{WT_sig}
\alias{WT_sig}
\title{Compute wavelet significance}
\usage{
WT_sig(data, mywt, alpha)
}
\arguments{
\item{data}{Numeric vector or matrix of input data.}

\item{mywt}{Wavelet object.}

\item{alpha}{Numeric, significance level.}
}
\value{
List with `sig1` and `sig2` numeric vectors.
}
\description{
Computes significance levels for a wavelet object.
}
\examples{
\dontrun{
sig <- WT_sig(matrix(rnorm(100), ncol=2), wt_obj, 0.05)
}
}
